+++ app/src/main/java/com/crystal2033/qrextractor/scanner_feature/data/remote/api/PersonApi.kt
@@ -7,5 +7,5 @@ import retrofit2.http.Path
 
 interface PersonApi {
     @GET("api/persons/{id}")
-    suspend fun getPerson(@Path(value = "id") id: Int) : Response<PersonDto?>
+    suspend fun getPerson(@Path(value = "id") id: Long) : Response<PersonDto?>
 }
