+++ app/src/main/java/com/crystal2033/qrextractor/scanner_feature/domain/model/Unknown.kt
@@ -5,7 +5,7 @@ class Unknown(val message: String) : QRScannableData {
         return DatabaseObjectTypes.UNKNOWN
     }
 
-    override fun getDatabaseID(): Int {
+    override fun getDatabaseID(): Long {
         return 0
     }
 }
